Additional changes to ServiceList

+++ intelligenceHomeApp/app/src/main/java/com/example/gaba/intelligence_homeapp/Service.java
@@ -3,14 +3,14 @@ package com.example.gaba.intelligence_homeapp;
 public class Service {
 
     private String name;
-    private int rate;
+    private Double rate;
 
     public Service(){
         name = null;
-        rate = 0;
+        rate = 0.0;
     }
 
-    public Service(String name, int rate){
+    public Service(String name, Double rate){
 
         this.name = name;
         this.rate = rate;
@@ -21,13 +21,13 @@ public class Service {
         return name;
     }
 
-    public int getRate() {
+    public Double getRate() {
         return rate;
     }
 
     public void setName(String name){this.name = name;}
 
-    public void setRate(int rate) {
+    public void setRate(Double rate) {
         this.rate = rate;
     }
